requirements.txt
asarPy==1.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
